+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
     - New group of algorithms which are used for evaluating whether the candidate is incorrect. In that case penalty is computed.
     - _Count_
         - Penalizer 'Count' is the basic penalizer used in ng_trajectory. It counts number of points outside the valid area.
+    - _Borderlines_
+        - Penalizer 'Borderlines' calculates minimum of maximum of the distances between misplaced points and their associated borderline (sets of points on the borders between two adjacent segments).
+        - **Warning** Currently, this works only for Matryoshka.
 - Segmentators
     - _Flood Fill_
         - Function `segmentDistance()` that computes distance from a point to a segment.